+++ docs/conf.py
@@ -42,7 +42,7 @@ primary_domain = 'py'
 default_role = 'py:obj'
 
 intersphinx_mapping = {
-    'python': ('http://python.readthedocs.io/en/latest/', None),
+    'python': ('https://python.readthedocs.io/en/latest/', None),
 }
 
 issues_github_path = 'marshmallow-code/marshmallow'
@@ -91,9 +91,9 @@ html_theme_options = {
     'warn_border': '#EEE',
     # Used to populate the useful-links.html template
     'extra_nav_links': OrderedDict([
-        ('marshmallow @ PyPI', 'http://pypi.python.org/pypi/marshmallow'),
-        ('marshmallow @ GitHub', 'http://github.com/marshmallow-code/marshmallow'),
-        ('Issue Tracker', 'http://github.com/marshmallow-code/marshmallow/issues'),
+        ('marshmallow @ PyPI', 'https://pypi.python.org/pypi/marshmallow'),
+        ('marshmallow @ GitHub', 'https://github.com/marshmallow-code/marshmallow'),
+        ('Issue Tracker', 'https://github.com/marshmallow-code/marshmallow/issues'),
     ]),
 }